requirements.txt
diffusers==0.33.0
dreamsim==0.2.1
imageio==2.33.0
lpips==0.1.4
matplotlib==3.10.1
nudenet==3.4.2
numpy==2.2.5
pandas==2.2.3
Pillow==11.2.1
Requests==2.32.3
safetensors==0.5.3
torch==2.4.0
torchvision==0.19.0
tqdm==4.66.5
transformers==4.48.0
